Refactor into structured files
