+++ lib/swagger/blocks/errors/declaration_error.rb
@@ -1,0 +1,8 @@
+module Swagger
+  module Blocks
+    module Errors
+      class DeclarationError < Error
+      end
+    end
+  end
+end
