+++ lib/swagger/blocks/errors/error.rb
@@ -1,0 +1,8 @@
+module Swagger
+  module Blocks
+    module Errors
+      class Error < StandardError
+      end
+    end
+  end
+end
