+++ lib/swagger/blocks/errors/not_found_error.rb
@@ -1,0 +1,8 @@
+module Swagger
+  module Blocks
+    module Errors
+      class NotFoundError < Error
+      end
+    end
+  end
+end
