+++ lib/swagger/blocks/errors/not_supported_error.rb
@@ -1,0 +1,8 @@
+module Swagger
+  module Blocks
+    module Errors
+      class NotSupportedError < Error
+      end
+    end
+  end
+end
