+++ lib/swagger/blocks/nodes/api_authorization_scope_node.rb
@@ -1,0 +1,10 @@
+module Swagger
+  module Blocks
+    module Nodes
+      # v1.2: NOTE: in the spec this is different than Resource Listing's scope object.
+      # v1.2: http://goo.gl/PvwUXj#5211-scope-object
+      class ApiAuthorizationScopeNode < Node
+      end
+    end
+  end
+end
