+++ lib/swagger/blocks/nodes/contact_node.rb
@@ -1,0 +1,9 @@
+module Swagger
+  module Blocks
+    module Nodes
+      # v2.0: https://github.com/swagger-api/swagger-spec/blob/master/versions/2.0.md#contact-object
+      class ContactNode < Node
+      end
+    end
+  end
+end
