+++ lib/swagger/blocks/nodes/example_node.rb
@@ -1,0 +1,9 @@
+module Swagger
+  module Blocks
+    module Nodes
+      # v2.0:
+      class ExampleNode < Node
+      end
+    end
+  end
+end
