+++ lib/swagger/blocks/nodes/external_docs_node.rb
@@ -1,0 +1,9 @@
+module Swagger
+  module Blocks
+    module Nodes
+      # v2.0: https://github.com/swagger-api/swagger-spec/blob/master/versions/2.0.md#externalDocumentationObject
+      class ExternalDocsNode < Node
+      end
+    end
+  end
+end
