+++ lib/swagger/blocks/nodes/header_node.rb
@@ -1,0 +1,12 @@
+module Swagger
+  module Blocks
+    module Nodes
+      # v2.0: https://github.com/swagger-api/swagger-spec/blob/master/versions/2.0.md#headerObject
+      class HeaderNode < Node
+        def items(inline_keys = nil, &block)
+          self.data[:items] = Swagger::Blocks::Nodes::ItemsNode.call(version: version, inline_keys: inline_keys, &block)
+        end
+      end
+    end
+  end
+end
