+++ lib/swagger/blocks/nodes/items_node.rb
@@ -1,0 +1,15 @@
+module Swagger
+  module Blocks
+    module Nodes
+      # v1.2:
+      # v2.0:
+      class ItemsNode < Node
+        def property(name, inline_keys = nil, &block)
+          self.data[:properties] ||= Swagger::Blocks::Nodes::PropertiesNode.new
+          self.data[:properties].version = version
+          self.data[:properties].property(name, inline_keys, &block)
+        end
+      end
+    end
+  end
+end
