+++ lib/swagger/blocks/nodes/login_endpoint_node.rb
@@ -1,0 +1,9 @@
+module Swagger
+  module Blocks
+    module Nodes
+      # v1.2: http://goo.gl/PvwUXj#5110-login-endpoint-object
+      class LoginEndpointNode < Node
+      end
+    end
+  end
+end
