+++ lib/swagger/blocks/nodes/property_node.rb
@@ -1,0 +1,18 @@
+module Swagger
+  module Blocks
+    module Nodes
+      # v1.2: http://goo.gl/PvwUXj#527-model-object
+      class PropertyNode < Node
+        def items(inline_keys = nil, &block)
+          self.data[:items] = Swagger::Blocks::Nodes::ItemsNode.call(version: version, inline_keys: inline_keys, &block)
+        end
+
+        def property(name, inline_keys = nil, &block)
+          self.data[:properties] ||= Swagger::Blocks::Nodes::PropertiesNode.new
+          self.data[:properties].version = version
+          self.data[:properties].property(name, inline_keys, &block)
+        end
+      end
+    end
+  end
+end
