+++ lib/swagger/blocks/nodes/resource_node.rb
@@ -1,0 +1,9 @@
+module Swagger
+  module Blocks
+    module Nodes
+      # v1.2: http://goo.gl/PvwUXj#512-resource-object
+      class ResourceNode < Node
+      end
+    end
+  end
+end
