+++ lib/swagger/blocks/nodes/scope_node.rb
@@ -1,0 +1,9 @@
+module Swagger
+  module Blocks
+    module Nodes
+      # v1.2: http://goo.gl/PvwUXj#516-scope-object
+      class ScopeNode < Node
+      end
+    end
+  end
+end
