+++ lib/swagger/blocks/nodes/token_endpoint_node.rb
@@ -1,0 +1,9 @@
+module Swagger
+  module Blocks
+    module Nodes
+      # v1.2: http://goo.gl/PvwUXj#5112-token-endpoint-object
+      class TokenEndpointNode < Node
+      end
+    end
+  end
+end
