+++ lib/swagger/blocks/nodes/token_request_endpoint_node.rb
@@ -1,0 +1,9 @@
+module Swagger
+  module Blocks
+    module Nodes
+      # v1.2: http://goo.gl/PvwUXj#5111-token-request-endpoint-object
+      class TokenRequestEndpointNode < Node
+      end
+    end
+  end
+end
