+++ lib/swagger/blocks/nodes/xml_node.rb
@@ -1,0 +1,9 @@
+module Swagger
+  module Blocks
+    module Nodes
+      # v2.0:
+      class XmlNode < Node
+      end
+    end
+  end
+end
